+++ addons/filamind_iot/models/__init__.py
@@ -1,4 +1,6 @@
 from . import iot_device_type
+from . import iot_channel
+from . import iot_keyboard_layout
 from . import iot_box
 from . import iot_device
 from . import iot_connection_log